requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asgiref
attrs
blinker
certifi
charset-normalizer
click
cursor
deprecation
distro
Django
django-cors-headers
filelock
frozenlist
fsspec
gotrue
h11
h2
hpack
httpcore
httpx
huggingface-hub
hyperframe
idna
inflect
Jinja2
jiter
joblib
MarkupSafe
more-itertools
mpmath
multidict
networkx
numpy
openai
opencv-python
packaging
pandas
pillow
postgrest
propcache
psycopg2-binary
pydantic
pydantic_core
python-dateutil
python-dotenv
pytz
PyYAML
pyzbar
RapidFuzz
realtime
regex
requests
safetensors
scikit-learn
scipy
setuptools
six
sniffio
sqlparse
storage3
StrEnum
supabase
supafunc
sympy
threadpoolctl
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
typeguard
typing_extensions
tzdata
urllib3
websockets
xgboost
yarl
celery
redis
djangorestframework
celery[redis]
channels
channels_redis